Refactor: Improve README and app structure
Update README with new features and improve app structure by moving components to dedicated folders.
